feat: добавление отображения фильмов
Пока не работают роутинги
