+++ Dockerfile
@@ -2,7 +2,7 @@ FROM python:3.13-alpine
 
 WORKDIR /app
 
-ENV FLASK_APP=app/main.py
+ENV FLASK_APP=app/run.py
 ENV FLASK_RUN_HOST=0.0.0.0
 
 COPY requirements.txt .
