+++ docker-compose.yml
@@ -5,9 +5,11 @@ services:
     build: .
     container_name: cinema-web
     ports:
-      - "5001:5000"
+      - "5000:5000"
     depends_on:
       - db
+    volumes:
+      - ./:/app
 
   db:
     image: mongo:6.0
